fix: Add .shlibs file, and update version to 0.1.1 (#21)
It seems the gnome-control-center binary ends up linking against the
SONAME symlink, so it should depend on a version that includes it. With
this, `dh_shlibdeps` generates such a dependency. This, or `.symbols`,
is also required by the debian policy manual, so it's probably best to
tend to follow the conventions for `.deb` packages.

This should not change the behavior of anything other than new packages
built against this library.